+++ apps/ledger-live-mobile/src/live-common-setup.ts
@@ -133,6 +133,8 @@ setSupportedCurrencies([
   "mina",
   "babylon",
   "canton_network",
+  "canton_network_devnet",
+  "canton_network_localnet",
 ]);
 
 if (Config.FORCE_PROVIDER && !isNaN(parseInt(Config.FORCE_PROVIDER, 10)))