+++ libs/device-core/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @ledgerhq/device-core
 
+## 0.6.1
+
+### Patch Changes
+
+- Updated dependencies [[`31dceca`](https://github.com/LedgerHQ/ledger-live/commit/31dcecad24abdaa8a77208e82532c47728c11180)]:
+  - @ledgerhq/types-live@6.82.0
+
+## 0.6.1-next.0
+
+### Patch Changes
+
+- Updated dependencies [[`31dceca`](https://github.com/LedgerHQ/ledger-live/commit/31dcecad24abdaa8a77208e82532c47728c11180)]:
+  - @ledgerhq/types-live@6.82.0-next.0
+
 ## 0.6.0
 
 ### Minor Changes