+++ libs/ledgerjs/packages/hw-app-celo/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @ledgerhq/hw-app-polkadot
 
+## 6.34.1
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @ledgerhq/hw-app-eth@6.45.16
+
+## 6.34.1-next.0
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @ledgerhq/hw-app-eth@6.45.16-next.0
+
 ## 6.34.0
 
 ### Minor Changes